+++ Sources/Validator.cs
@@ -45,28 +45,26 @@ public class Validator {
 /// <summary>
 /// A validator that compares a value to another reference value.
 /// </summary>
-/// <typeparam name="T">The type of the comparison value.</typeparam>
-public class ComparisonValidator<T>: Validator {
+public class ComparisonValidator: Validator {
 
 	/// <summary>
 	/// The comparison value.
 	/// </summary>
-	public required T Value { get; init; }
+	public required object? Value { get; init; }
 }
 
 /// <summary>
 /// A validator that ensures that a value falls within a specified range.
 /// </summary>
-/// <typeparam name="T">The type of comparison values.</typeparam>
-public class RangeValidator<T>: Validator {
+public class RangeValidator: Validator {
 
 	/// <summary>
 	/// The lower bound.
 	/// </summary>
-	public required T LowerBound { get; init; }
+	public required object LowerBound { get; init; }
 
 	/// <summary>
 	/// The upper bound.
 	/// </summary>
-	public required T UpperBound { get; init; }
+	public required object UpperBound { get; init; }
 }