+++ ChangeLog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version [0.2.0](https://github.com/CedX/Validation.ps1/compare/v0.1.1...v0.2.0)
-- Added the `New-ValidatorIn` `New-ValidatorNotIn` and `New-ValidatorCreditCard` cmdlets.
+- Added the `New-ValidatorIn`, `New-ValidatorNotIn` and `New-ValidatorCreditCard` cmdlets.
 
 ## Version [0.1.1](https://github.com/CedX/Validation.ps1/compare/v0.1.0...v0.1.1)
 - Fixed a packaging issue.